feat: remove redundant code
